feat: сделан фнкционал меняющегося интерфейса в хедере при входе-выходе
