+++ src/components/header/header.tsx
@@ -1,19 +1,14 @@
-import { Link, NavLink } from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 import { AppRoute } from '../../const';
-import { useAppDispatch } from '../../hooks/useAppDispatch/useAppDispatch';
-import { logout } from '../../store/api-actions';
-
-const getStyleForNavLink = ({ isActive }: { isActive: boolean }) =>
-  isActive
-    ? {
-      cursor: 'default',
-    }
-    : {
-      cursor: 'pointer',
-    };
+import * as selectors from '../../store/selectors';
+import { useAppSelector } from '../../hooks/useAppSelector/useAppSelector';
+import { AuthorizationStatus } from '../../const';
+import { getStyleForNavLink } from '../../utils';
+import { AuthElement } from '../log-elements/auth-element/auth-element';
+import { NoAuthElement } from '../log-elements/no-auth-element/no-auth-element';
 
 export function Header() {
-  const dispatch = useAppDispatch();
+  const authStatus = useAppSelector(selectors.authorizationStatus);
 
   return (
     <header className="header">
@@ -30,35 +25,8 @@ export function Header() {
               />
             </NavLink>
           </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              <li className="header__nav-item user">
-                <NavLink
-                  className="header__nav-link header__nav-link--profile"
-                  to={AppRoute.Favorites}
-                  style={getStyleForNavLink}
-                >
-                  <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                  <span className="header__user-name user__name">
-                    Oliver.conner@gmail.com
-                  </span>
-                  <span className="header__favorite-count">3</span>
-                </NavLink>
-              </li>
-              <li className="header__nav-item">
-                <Link
-                  className="header__nav-link"
-                  to={AppRoute.Root}
-                  onClick={(evt) => {
-                    evt.preventDefault();
-                    dispatch(logout());
-                  }}
-                >
-                  <span className="header__signout">Sign out</span>
-                </Link>
-              </li>
-            </ul>
-          </nav>
+          {authStatus === AuthorizationStatus.Auth ? <AuthElement /> : <NoAuthElement />}
+
         </div>
       </div>
     </header>
