+++ src/components/log-elements/no-auth-element/no-auth-element.tsx
@@ -1,0 +1,21 @@
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../../const';
+
+
+export function NoAuthElement() {
+  return (
+    <nav className="header__nav">
+      <ul className="header__nav-list">
+        <li className="header__nav-item user">
+          <Link
+            className="header__nav-link"
+            to={AppRoute.Login}
+          >
+            <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+            <span className="header__login">Sign in</span>
+          </Link>
+        </li>
+      </ul>
+    </nav>
+  );
+}
